make max_prefix an u_int32_t instead of an u_long and change its meaning as in 0 means no limit insytead of setting it to ULONG_MAX for no limit

+++ bgpd.h
@@ -131,7 +131,7 @@ struct peer_config {
 	char			 descr[PEER_DESCR_LEN];
 	struct bgpd_addr	 remote_addr;
 	struct bgpd_addr	 local_addr;
-	u_long			 max_prefix;
+	u_int32_t		 max_prefix;
 	u_int16_t		 remote_as;
 	u_int8_t		 ebgp;		/* 1 = ebgp, 0 = ibgp */
 	u_int8_t		 distance;	/* 1 = direct, >1 = multihop */
